// rugchecker/model/result/metadata-check.ts
export default class MetadataCheckResult {
  address: string;
  name: string;
  symbol: string;
  imageUrl: string;
  description: string;
  website: string;
  telegram: string;
  twitter: string;
  isMutable: boolean;
  isMintable: boolean;
  isFreezable: boolean;
  isPumpFun: boolean;
  isValid: boolean;
}